update triple client and server api
