+++ client/options.go
@@ -18,6 +18,8 @@
 package client
 
 import (
+	"dubbo.apache.org/dubbo-go/v3/common/dubboutil"
+	"reflect"
 	"strconv"
 	"time"
 )
@@ -39,11 +41,9 @@ import (
 )
 
 type ReferenceOptions struct {
-	Reference   *global.ReferenceConfig
-	Application *global.ApplicationConfig
-	Registries  map[string]*global.RegistryConfig
-	Shutdown    *global.ShutdownConfig
-	cliOpts     *ClientOptions
+	Reference  *global.ReferenceConfig
+	cliOpts    *ClientOptions
+	Registries map[string]*global.RegistryConfig
 
 	pxy          *proxy.Proxy
 	id           string
@@ -59,9 +59,7 @@ type ReferenceOptions struct {
 
 func defaultReferenceOptions() *ReferenceOptions {
 	return &ReferenceOptions{
-		Application: global.DefaultApplicationConfig(),
-		Reference:   global.DefaultReferenceConfig(),
-		Shutdown:    global.DefaultShutdownConfig(),
+		Reference: global.DefaultReferenceConfig(),
 	}
 }
 
@@ -73,7 +71,8 @@ func (refOpts *ReferenceOptions) init(cli *Client, opts ...ReferenceOption) erro
 		return err
 	}
 
-	refOpts.setRefWithCon(cli.cliOpts.Consumer)
+	refOpts.cliOpts = cli.cliOpts
+	dubboutil.CopyFields(reflect.ValueOf(refOpts.cliOpts.Consumer).Elem(), reflect.ValueOf(refOpts.Reference).Elem())
 
 	ref := refOpts.Reference
 
@@ -90,7 +89,7 @@ func (refOpts *ReferenceOptions) init(cli *Client, opts ...ReferenceOption) erro
 	}
 
 	// init application
-	application := refOpts.Application
+	application := refOpts.cliOpts.Application
 	if application != nil {
 		refOpts.applicationCompat = compatApplicationConfig(application)
 		if err := refOpts.applicationCompat.Init(); err != nil {
@@ -116,9 +115,6 @@ func (refOpts *ReferenceOptions) init(cli *Client, opts ...ReferenceOption) erro
 		emptyRegIDsFlag = true
 	}
 	regs := refOpts.Registries
-	if regs == nil {
-		regs = cli.cliOpts.Registries
-	}
 	if regs != nil {
 		refOpts.registriesCompat = make(map[string]*config.RegistryConfig)
 		for key, reg := range regs {
@@ -134,13 +130,9 @@ func (refOpts *ReferenceOptions) init(cli *Client, opts ...ReferenceOption) erro
 	ref.RegistryIDs = commonCfg.TranslateIds(ref.RegistryIDs)
 
 	// init graceful_shutdown
-	graceful_shutdown.Init(graceful_shutdown.SetShutdown_Config(refOpts.Shutdown))
+	graceful_shutdown.Init(graceful_shutdown.SetShutdown_Config(refOpts.cliOpts.Shutdown))
 
 	return commonCfg.Verify(refOpts)
-}
-
-func (refOpts *ReferenceOptions) setRefWithCon(consumer *global.ConsumerConfig) {
-	// set fields of refOpts.Reference with consumer as default value
 }
 
 type ReferenceOption func(*ReferenceOptions)
@@ -179,19 +171,11 @@ func WithRegistryIDs(registryIDs []string) ReferenceOption {
 func WithRegistry(opts ...registry.Option) ReferenceOption {
 	regOpts := registry.NewOptions(opts...)
 
-	return func(cliOpts *ReferenceOptions) {
-		if cliOpts.Registries == nil {
-			cliOpts.Registries = make(map[string]*global.RegistryConfig)
+	return func(refOpts *ReferenceOptions) {
+		if refOpts.Registries == nil {
+			refOpts.Registries = make(map[string]*global.RegistryConfig)
 		}
-		cliOpts.Registries[regOpts.ID] = regOpts.Registry
-	}
-}
-
-func WithShutdown(opts ...graceful_shutdown.Option) ReferenceOption {
-	sdOpts := graceful_shutdown.NewOptions(opts...)
-
-	return func(cliOpts *ReferenceOptions) {
-		cliOpts.Shutdown = sdOpts.Shutdown
+		refOpts.Registries[regOpts.ID] = regOpts.Registry
 	}
 }
 
@@ -402,20 +386,305 @@ func SetRegistries(regs map[string]*global.RegistryConfig) ReferenceOption {
 	}
 }
 
-func SetApplication(application *global.ApplicationConfig) ReferenceOption {
-	return func(opts *ReferenceOptions) {
-		opts.Application = application
-	}
-}
-
 func SetReference(reference *global.ReferenceConfig) ReferenceOption {
 	return func(opts *ReferenceOptions) {
 		opts.Reference = reference
 	}
 }
 
-func SetShutdown(shutdown *global.ShutdownConfig) ReferenceOption {
-	return func(opts *ReferenceOptions) {
+type ClientOptions struct {
+	Consumer    *global.ConsumerConfig
+	Application *global.ApplicationConfig
+	Registries  map[string]*global.RegistryConfig
+	Shutdown    *global.ShutdownConfig
+}
+
+func defaultClientOptions() *ClientOptions {
+	return &ClientOptions{
+		Consumer:    global.DefaultConsumerConfig(),
+		Registries:  make(map[string]*global.RegistryConfig),
+		Application: global.DefaultApplicationConfig(),
+		Shutdown:    global.DefaultShutdownConfig(),
+	}
+}
+
+func (cliOpts *ClientOptions) init(opts ...ClientOption) error {
+	for _, opt := range opts {
+		opt(cliOpts)
+	}
+	if err := defaults.Set(cliOpts); err != nil {
+		return err
+	}
+	return nil
+}
+
+type ClientOption func(*ClientOptions)
+
+func WithClientCheck() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Check = true
+	}
+}
+
+func WithClientURL(url string) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.URL = url
+	}
+}
+
+// todo(DMwangnima): change Filter Option like Cluster and LoadBalance
+func WithClientFilter(filter string) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Filter = filter
+	}
+}
+
+// todo(DMwangnima): think about a more ideal configuration style
+func WithClientRegistryIDs(registryIDs []string) ClientOption {
+	return func(opts *ClientOptions) {
+		if len(registryIDs) > 0 {
+			opts.Consumer.RegistryIDs = registryIDs
+		}
+	}
+}
+
+func WithClientRegistry(opts ...registry.Option) ClientOption {
+	regOpts := registry.NewOptions(opts...)
+
+	return func(cliOpts *ClientOptions) {
+		if cliOpts.Registries == nil {
+			cliOpts.Registries = make(map[string]*global.RegistryConfig)
+		}
+		cliOpts.Registries[regOpts.ID] = regOpts.Registry
+	}
+}
+
+//func WithClientShutdown(opts ...graceful_shutdown.Option) ClientOption {
+//	sdOpts := graceful_shutdown.NewOptions(opts...)
+//
+//	return func(cliOpts *ClientOptions) {
+//		cliOpts.Shutdown = sdOpts.Shutdown
+//	}
+//}
+
+// ========== Cluster Strategy ==========
+
+func WithClientClusterAvailable() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Cluster = constant.ClusterKeyAvailable
+	}
+}
+
+func WithClientClusterBroadcast() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Cluster = constant.ClusterKeyBroadcast
+	}
+}
+
+func WithClientClusterFailBack() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Cluster = constant.ClusterKeyFailback
+	}
+}
+
+func WithClientClusterFailFast() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Cluster = constant.ClusterKeyFailfast
+	}
+}
+
+func WithClientClusterFailOver() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Cluster = constant.ClusterKeyFailover
+	}
+}
+
+func WithClientClusterFailSafe() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Cluster = constant.ClusterKeyFailsafe
+	}
+}
+
+func WithClientClusterForking() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Cluster = constant.ClusterKeyForking
+	}
+}
+
+func WithClientClusterZoneAware() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Cluster = constant.ClusterKeyZoneAware
+	}
+}
+
+func WithClientClusterAdaptiveService() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Cluster = constant.ClusterKeyAdaptiveService
+	}
+}
+
+// ========== LoadBalance Strategy ==========
+
+func WithClientLoadBalanceConsistentHashing() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Loadbalance = constant.LoadBalanceKeyConsistentHashing
+	}
+}
+
+func WithClientLoadBalanceLeastActive() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Loadbalance = constant.LoadBalanceKeyLeastActive
+	}
+}
+
+func WithClientLoadBalanceRandom() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Loadbalance = constant.LoadBalanceKeyRandom
+	}
+}
+
+func WithClientLoadBalanceRoundRobin() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Loadbalance = constant.LoadBalanceKeyRoundRobin
+	}
+}
+
+func WithClientLoadBalanceP2C() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Loadbalance = constant.LoadBalanceKeyP2C
+	}
+}
+
+func WithClientLoadBalanceXDSRingHash() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Loadbalance = constant.LoadBalanceKeyLeastActive
+	}
+}
+
+func WithClientRetries(retries int) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Retries = strconv.Itoa(retries)
+	}
+}
+
+func WithClientGroup(group string) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Group = group
+	}
+}
+
+func WithClientVersion(version string) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Version = version
+	}
+}
+
+func WithClientJSON() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Serialization = constant.JSONSerialization
+	}
+}
+
+func WithClientProvidedBy(providedBy string) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.ProvidedBy = providedBy
+	}
+}
+
+// todo(DMwangnima): implement this functionality
+//func WithAsync() ClientOption {
+//	return func(opts *ClientOptions) {
+//		opts.Consumer.Async = true
+//	}
+//}
+
+func WithClientParams(params map[string]string) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Params = params
+	}
+}
+
+// todo(DMwangnima): implement this functionality
+//func WithClientGeneric(generic bool) ClientOption {
+//	return func(opts *ClientOptions) {
+//		if generic {
+//			opts.Consumer.Generic = "true"
+//		} else {
+//			opts.Consumer.Generic = "false"
+//		}
+//	}
+//}
+
+func WithClientSticky(sticky bool) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Sticky = sticky
+	}
+}
+
+// ========== Protocol to consume ==========
+
+func WithClientProtocolDubbo() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Protocol = constant.Dubbo
+	}
+}
+
+func WithClientProtocolTriple() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Protocol = "tri"
+	}
+}
+
+func WithClientProtocolJsonRPC() ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Protocol = "jsonrpc"
+	}
+}
+
+func WithClientProtocol(protocol string) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.Protocol = protocol
+	}
+}
+
+func WithClientRequestTimeout(timeout time.Duration) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.RequestTimeout = timeout.String()
+	}
+}
+
+func WithClientForce(force bool) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.ForceTag = force
+	}
+}
+
+func WithClientMeshProviderPort(port int) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer.MeshProviderPort = port
+	}
+}
+
+func SetClientRegistries(regs map[string]*global.RegistryConfig) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Registries = regs
+	}
+}
+
+func SetApplication(application *global.ApplicationConfig) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Application = application
+	}
+}
+
+func SetConsumer(consumer *global.ConsumerConfig) ClientOption {
+	return func(opts *ClientOptions) {
+		opts.Consumer = consumer
+	}
+}
+
+func SetShutdown(shutdown *global.ShutdownConfig) ClientOption {
+	return func(opts *ClientOptions) {
 		opts.Shutdown = shutdown
 	}
 }
