+++ protocol/triple/internal/client/health_client/main.go
@@ -33,7 +33,7 @@ import (
 
 func main() {
 	cli, err := client.NewClient(
-		client.WithURL("tri://127.0.0.1:20000"),
+		client.WithClientURL("tri://127.0.0.1:20000"),
 	)
 	if err != nil {
 		panic(err)
